+++ packages/store/src/module.ts
@@ -2,7 +2,7 @@ import { NgModule, ModuleWithProviders, Optional, Inject } from '@angular/core';
 
 import { ROOT_STATE_TOKEN, FEATURE_STATE_TOKEN } from './symbols';
 import { StateFactory } from './state-factory';
-import { Actions } from './actions-stream';
+import { Actions, InternalActions } from './actions-stream';
 import { InternalActionCompletions, ActionCompletions } from './actions-completion-stream';
 import { Store } from './store';
 import { SelectFactory } from './select';
@@ -89,6 +89,7 @@ export class NgxsModule {
       providers: [
         StateFactory,
         Actions,
+        InternalActions,
         InternalActionCompletions,
         ActionCompletions,
         Store,