+++ packages/store/src/store.ts
@@ -1,26 +1,24 @@
 import { Injectable, ErrorHandler } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Subscription } from 'rxjs/Subscription';
-import { distinctUntilChanged, catchError, take, shareReplay, tap } from 'rxjs/operators';
+import { distinctUntilChanged, catchError, take, shareReplay } from 'rxjs/operators';
 import { forkJoin } from 'rxjs/observable/forkJoin';
 import { map } from 'rxjs/operators/map';
 import { of } from 'rxjs/observable/of';
 
 import { compose } from './compose';
-import { Actions } from './actions-stream';
+import { InternalActions } from './actions-stream';
 import { StateFactory } from './state-factory';
 import { StateStream } from './state-stream';
 import { PluginManager } from './plugin-manager';
 import { fastPropGetter } from './internals';
 import { META_KEY } from './symbols';
-import { InternalActionCompletions } from './actions-completion-stream';
 
 @Injectable()
 export class Store {
   constructor(
     private _errorHandler: ErrorHandler,
-    private _actions: Actions,
-    private _actionCompletions: InternalActionCompletions,
+    private _actions: InternalActions,
     private _storeFactory: StateFactory,
     private _pluginManager: PluginManager,
     private _stateStream: StateStream
@@ -128,7 +126,7 @@ export class Store {
             this._actions,
             action
           )
-          .pipe(tap(() => this._actionCompletions.next(nextAction)), map(() => this._stateStream.getValue()));
+          .pipe(map(() => this._stateStream.getValue()));
       }
     ])(prevState, action) as Observable<any>).pipe(shareReplay());
   }