+++ packages/store/tests/action.spec.ts
@@ -5,7 +5,6 @@ import { timer } from 'rxjs/observable/timer';
 import { TestBed } from '@angular/core/testing';
 import { NgxsModule } from '../src/module';
 import { Store } from '../src/store';
-import { ActionCompletions } from '../src/actions-completion-stream';
 import { Actions } from '../src/actions-stream';
 import { tap } from 'rxjs/operators';
 
@@ -59,37 +58,6 @@ describe('Actions', () => {
 
     actions.subscribe(action => {
       expect(happened).toBeFalsy();
-    });
-
-    store.dispatch(new Action1());
-  });
-});
-
-describe('Action Completions', () => {
-  it('basic', () => {
-    let happened = false;
-
-    class Action1 {
-      static type = 'ACTION 1';
-    }
-
-    @State({ name: 'Bar' })
-    class BarStore {
-      @Action(Action1)
-      bar() {
-        return timer(10).pipe(tap(() => (happened = true)));
-      }
-    }
-
-    TestBed.configureTestingModule({
-      imports: [NgxsModule.forRoot([BarStore])]
-    });
-
-    const store = TestBed.get(Store);
-    const actions = TestBed.get(ActionCompletions);
-
-    actions.subscribe(action => {
-      expect(happened).toBeTruthy();
     });
 
     store.dispatch(new Action1());